+++ lib/hpsc/link-mbox.h
@@ -13,6 +13,6 @@
 // To claim as client: set server to 0 and set client to non-zero ID
 struct link *link_mbox_connect(const char *name, struct hpsc_mbox *mbox,
                                unsigned idx_from, unsigned idx_to,
-                               uint32_t server, uint32_t client);
+                               uint8_t server, uint8_t client);
 
 #endif // LINK_MBOX_H
